requirements.txt
# Production dashboard dependencies.
# Kept lightweight for Streamlit Cloud/Render free-tier deployment.

pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.0

streamlit>=1.32.0
plotly>=5.18.0
kaleido>=0.2.1

matplotlib>=3.8.0
seaborn>=0.13.0

scikit-learn>=1.4.0

requests>=2.31.0
python-dotenv>=1.0.0

python-docx>=1.1.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
